Move Pill component to common directrory
