+++ assets/js/components/InstanceOverview/InstanceOverview.jsx
@@ -6,7 +6,7 @@ import HealthIcon from '@components/Health';
 import { Features } from '@components/SapSystemDetails';
 import HostLink from '@components/HostLink';
 import ClusterLink from '@pages/ClusterDetails/ClusterLink';
-import Pill from '@components/Pill';
+import Pill from '@common/Pill';
 import CleanUpButton from '@components/CleanUpButton';
 import Tooltip from '@common/Tooltip';
