+++ assets/js/trento.jsx
@@ -12,7 +12,7 @@ import Layout from '@components/Layout';
 import Home from '@components/Home';
 import AboutPage from '@pages/AboutPage';
 import HostsList from '@components/HostsList';
-import ClustersList from '@components/ClustersList';
+import ClustersList from '@pages/ClusterDetails/ClustersList';
 import ClusterDetailsPage from '@pages/ClusterDetails';
 import ClusterSettingsPage from '@components/ClusterSettingsPage';
 import { ExecutionResultsPage } from '@components/ExecutionResults';
